Add pictures, reformatting, editing.
Added
   attachments/202508051556-000.jpeg
   attachments/202508051638-001.jpeg
   attachments/202508061209-002.jpeg
   attachments/202508071248-001.jpeg
   attachments/202508080656-000.jpeg
   attachments/202508080840-000.jpeg
   attachments/202508081609-000.jpeg
and added links to Hunters-Island-2025.md

Rounded all coordinates to 5 decimal places which should give better
than 20 meter precision. (Apple is silly to give so many.)

Started adding links to each day at the top.

Because of mosquitos, the log for 2025-08-08 was mostly part of the
log for the next day. Started to move the happenings of 2025-08-08
under their own heading.

That's when I began to doubt some of what I'd written, so I need to
spend time checking dates. That is, I don't trust the level 2 headings.